+++ HackerRank/Python/py-division.py
@@ -1,0 +1,15 @@
+# =====================================================================
+# Author:       James Walker (GothicAnatomist)
+# Create date:  03/07/2021
+# Site:         HackerRank
+# URL:          https://www.hackerrank.com/challenges/python-division/problem
+# Description:  Solution to "Python: Division"
+# Licence:      CC BY 4.0
+# =====================================================================
+
+if __name__ == '__main__':
+    a = int(input())
+    b = int(input())
+
+    print(a // b)  # int division
+    print(a / b)  # float division